(fix|COS-1952): Make contract link a hyperlink
