+++ packages/apps/spaces/ui/form/UrlInput/FormUrlInput.tsx
@@ -1,4 +1,7 @@
+import React from 'react';
 import { useField } from 'react-inverted-form';
+
+import { FormLabelProps } from '@chakra-ui/react';
 
 import { InputProps } from '@ui/form/Input';
 
@@ -7,6 +10,11 @@ import { UrlInput } from './UrlInput';
 interface FormUrlInputProps extends InputProps {
   name: string;
   formId: string;
+  label?: string;
+  isLabelVisible?: boolean;
+  labelProps?: FormLabelProps;
+  leftElement?: React.ReactNode;
+  rightElement?: React.ReactNode;
 }
 
 export const FormUrlInput = ({ name, formId, ...props }: FormUrlInputProps) => {
